+++ src/Contracts/MessengerAuthenticatable.php
@@ -4,12 +4,15 @@ namespace NettSite\Messenger\Contracts;
 
 use Illuminate\Contracts\Auth\Authenticatable;
 use Illuminate\Database\Eloquent\Relations\MorphMany;
+use Illuminate\Database\Eloquent\Relations\MorphOne;
 use Illuminate\Database\Eloquent\Relations\MorphToMany;
 use NettSite\Messenger\Models\DeviceToken;
 use NettSite\Messenger\Models\Message;
 
 interface MessengerAuthenticatable extends Authenticatable
 {
+    public function messengerEnrollment(): MorphOne;
+
     public function messageReceipts(): MorphMany;
 
     public function groups(): MorphToMany;