+++ src/Enums/RegistrationMode.php
@@ -5,5 +5,6 @@ namespace NettSite\Messenger\Enums;
 enum RegistrationMode: string
 {
     case Open = 'open';
+    case Approval = 'approval';
     case Closed = 'closed';
 }