+++ src/Traits/HasMessenger.php
@@ -3,14 +3,21 @@
 namespace NettSite\Messenger\Traits;
 
 use Illuminate\Database\Eloquent\Relations\MorphMany;
+use Illuminate\Database\Eloquent\Relations\MorphOne;
 use Illuminate\Database\Eloquent\Relations\MorphToMany;
 use NettSite\Messenger\Models\DeviceToken;
 use NettSite\Messenger\Models\Group;
 use NettSite\Messenger\Models\Message;
 use NettSite\Messenger\Models\MessageReceipt;
+use NettSite\Messenger\Models\MessengerEnrollment;
 
 trait HasMessenger
 {
+    public function messengerEnrollment(): MorphOne
+    {
+        return $this->morphOne(MessengerEnrollment::class, 'user');
+    }
+
     public function deviceTokens(): MorphMany
     {
         return $this->morphMany(DeviceToken::class, 'user');